Add method protection in Payment service

+++ DormService/Services/Payment/Payment.API/Controllers/DebtsController.cs
@@ -1,9 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Payment.Common.Entities;
 using Payment.Common.Repository;
+using System.Security.Claims;
 
 namespace Payment.API.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/v1/[controller]")]
     public class DebtsController : ControllerBase
@@ -15,10 +18,17 @@ namespace Payment.API.Controllers
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
         }
 
+
+        [Authorize(Roles = "Student, Administrator")]
         [HttpGet("{studentID}", Name = "GetStudentDebts")]
         [ProducesResponseType(typeof(StudentDebts), StatusCodes.Status200OK)]
         public async Task<ActionResult<StudentDebts>> GetStudentDebts(string studentID)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != studentID)
+            {
+                return Forbid();
+            }
+
             var studentDebts = await _repository.GetStudentDebts(studentID);
             if (studentDebts == null)
             {
@@ -27,28 +37,48 @@ namespace Payment.API.Controllers
             return Ok(studentDebts);
         }
 
+
+        [Authorize(Roles = "Student,Administrator")]
         [HttpPut]
         [ProducesResponseType(typeof(StudentDebts), StatusCodes.Status200OK)]
         public async Task<ActionResult> UpdateStudentDebt([FromBody] StudentDebts studentDebts)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != studentDebts.studentID)
+            {
+                return Forbid();
+            }
+
             return Ok(await _repository.UpdateStudentDebt(studentDebts));
         }
 
+
+        [Authorize(Roles = "Administrator")]
         [HttpPost]
         [ProducesResponseType(typeof(IEnumerable<StudentDebts>), StatusCodes.Status201Created)]
         public async Task<ActionResult<StudentDebts>> CreateStudent([FromBody] StudentDebts studentDebts)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != studentDebts.studentID)
+            {
+                return Forbid();
+            }
+
             await _repository.CreateNewStudent(studentDebts);
 
             return CreatedAtRoute("GetStudentDebts", new { studentID = studentDebts.studentID }, studentDebts);
         }
 
+
+        [Authorize(Roles = "Administrator")]
         [HttpDelete("{studentID}")]
         [ProducesResponseType(typeof(StudentDebts), StatusCodes.Status200OK)]
         public async Task<IActionResult> DeleteStudent(string studentID)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != studentID)
+            {
+                return Forbid();
+            }
+
             return Ok(await _repository.DeleteStudent(studentID));
         }
-
     }
 }
